refactor(ucmc-web): extract approved directory into ApprovedTab component
Move the search/filter/sort/list+grid/pagination/role-assign body out of
`routes/members.index.tsx` into `features/members/components/approved-tab.tsx`.
URL state stays owned by the route and threads through props. Mirrors
the shape now used by PendingTab/UnclaimedTab/LifecycleTab so all four
status views are siblings of the same kind. No user-visible change.
